alternatives/server-django/requirements-prod.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements-prod.in
#
asgiref==3.5.2
    # via django
attrs==21.2.0
    # via jsonschema
azure-core==1.17.0
    # via
    #   azure-eventgrid
    #   azure-storage-blob
    #   azure-storage-queue
azure-eventgrid==4.5.0
    # via -r requirements.in
azure-storage-blob==12.8.1
    # via -r requirements.in
azure-storage-queue==12.1.6
    # via -r requirements.in
boto3==1.16.53
    # via -r requirements.in
botocore==1.19.53
    # via
    #   boto3
    #   s3transfer
cachetools==4.2.2
    # via google-auth
certifi==2020.12.5
    # via
    #   minio
    #   msrest
    #   requests
cffi==1.14.6
    # via
    #   cryptography
    #   google-crc32c
charset-normalizer==2.0.4
    # via requests
colour==0.1.5
    # via styleframe
cryptography==3.4.7
    # via
    #   azure-storage-blob
    #   azure-storage-queue
    #   msal
    #   pyjwt
django==4.0.4
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-extensions
    #   djangorestframework
django-cors-headers==3.12.0
    # via -r requirements.in
django-extensions==3.1.5
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
google-api-core==2.0.0
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.0.1
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.0.0
    # via google-cloud-storage
google-cloud-storage==1.42.0
    # via -r requirements.in
google-crc32c==1.1.2
    # via google-resumable-media
google-resumable-media==2.0.0
    # via google-cloud-storage
googleapis-common-protos==1.53.0
    # via google-api-core
greenlet==1.1.1
    # via sqlalchemy
idna==3.2
    # via requests
isodate==0.6.0
    # via msrest
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
json-log-formatter==0.5.1
    # via -r requirements.in
jsonschema==3.2.0
    # via styleframe
minio==7.1.0
    # via -r requirements.in
msal==1.13.0
    # via -r requirements.in
msrest==0.6.21
    # via
    #   azure-eventgrid
    #   azure-storage-blob
    #   azure-storage-queue
numpy==1.21.2
    # via pandas
oauthlib==3.1.1
    # via requests-oauthlib
openpyxl==3.0.7
    # via
    #   -r requirements.in
    #   styleframe
pandas==1.3.2
    # via
    #   -r requirements.in
    #   styleframe
protobuf==3.17.3
    # via
    #   google-api-core
    #   googleapis-common-protos
psycopg2-binary==2.8.6
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.20
    # via cffi
pydantic==1.7.3
    # via -r requirements.in
pyjwt[crypto]==2.1.0
    # via msal
pyrsistent==0.18.0
    # via jsonschema
python-dateutil==2.8.1
    # via
    #   botocore
    #   pandas
pytz==2021.1
    # via
    #   djangorestframework
    #   pandas
requests==2.26.0
    # via
    #   -r requirements.in
    #   azure-core
    #   google-api-core
    #   google-cloud-storage
    #   msal
    #   msrest
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via msrest
rsa==4.7.2
    # via google-auth
s3transfer==0.3.4
    # via boto3
six==1.15.0
    # via
    #   azure-core
    #   isodate
    #   jsonschema
    #   protobuf
    #   python-dateutil
sqlalchemy==1.4.23
    # via -r requirements.in
sqlparse==0.4.2
    # via django
styleframe==4.0.0
    # via -r requirements.in
urllib3==1.26.2
    # via
    #   botocore
    #   minio
    #   requests
uwsgi==2.0.19.1
    # via -r requirements-prod.in
uwsgidecorators==1.1.0
    # via -r requirements-prod.in
werkzeug==2.1.2
    # via -r requirements.in
xlrd==1.2.0
    # via styleframe

# The following packages are considered to be unsafe in a requirements file:
# setuptools
